Updated to support receiving position messages

+++ Player.cs
@@ -4,8 +4,8 @@ namespace TestMmoServer
     {
         public string Id { get; set; }
 
-        public int X { get; set; }
+        public float X { get; set; }
 
-        public int Y { get; set; }
+        public float Y { get; set; }
     }
 }
